UPSTREAM: selftests, arm64: add a selftest for passing tagged pointers to kernel
(Upstream commit 9ce1263033cd2ad393e2ff0df4a1c4ab4992c9df).

This patch is a part of a series that extends kernel ABI to allow to pass
tagged user pointers (with the top byte set to something else other than
0x00) as syscall arguments.

This patch adds a simple test, that calls the uname syscall with a
tagged user pointer as an argument. Without the kernel accepting tagged
user pointers the test fails with EFAULT.

Reviewed-by: Catalin Marinas <catalin.marinas@arm.com>
Acked-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Andrey Konovalov <andreyknvl@google.com>
Signed-off-by: Will Deacon <will@kernel.org>
Signed-off-by: Andrey Konovalov <andreyknvl@google.com>
Bug: 135692346
Change-Id: I150694bf72706b5dd39939e38435c3aebf6ab2b3
Signed-off-by: Yousef Algadri <yusufgadrie@gmail.com>
Signed-off-by: Sudeep Duhoon <rishi.gothic@gmail.com>